feat(easter-egg): Implement interactive bug character
Adds a hidden, interactive bug character to the application that can be triggered and will wander around the screen. Includes new CSS animations for bug movement and behaviors, as well as updated TypeScript for state management of the bug's appearance and actions. The bug can be dismissed and will periodically respawn.